emerge-gitclone: use correct branch name format for commit
So far emerge-gitclone always checked out an exact commit in SHA1
format, which was given by the default manifest. Now that the
manifest started to specify the `revision` tag also as a branch name,
`emerge-gitclone` started to fail, as it cannot check out the branch
name given as `refs/heads/x`. The branch `x` was not checked out at the
moment.

So we need to make emerge-gitclone check out its exact remote branch
name, like `refs/remotes/origin/x`, in case of the branch name being
specified as a revision.