emerge-gitclone: fetch correct commit in case of non-release
When the given release string is for non-release like
"2022.02.02+dev-flatcar-master-4742", we should fetch release.xml from
a correct commit from
https://raw.githubusercontent.com/kinvolk/manifest-builds/dev-flatcar-master-4742/dev-flatcar-master-4742.xml
Without that, as the default branch contains invalid source code
that was deprecated many years ago, the build could sometimes fail, e.g.
when trying to build perl 5.26.2 with gcc 10.