emerge-gitclone: add support for pure git refs (dev images)
This change adds support for checking out git refs. It is useful for
using dev containers on development OS images.

For releases and for nightly builds, the OS version always corresponds
to a tag in the scripts repo. emerge-gitclone uses the OS version to
check out that tag. this works well for versions like e.g.
   alpha-3549.0.0
   3552.0.0+nightly-20230323-2100
because corresponding tags exist in the scripts repository.

However, a development OS image build (e.g. built on a dev machine, or
via CI) might use a version like e.g.
   alpha-3549.0.0-3-g7f21ab30
   3552.0.0+nightly-20230323-2100-4-g5d72f4ee
This version is generated by the SDK container using
   "git describe --tags"
in
   https://github.com/flatcar/scripts/blob/main/sdk_lib/sdk_container_common.sh#L49
.
emerge-gitclone will fail in that case because there is no tag of that
name in the scripts repo.

This change detects git refs appended to OS release versions. It
extracts the git ref and uses the ref instead of a tag for checking out
the correct scripts repo version.