Use FETCH_HEAD to fetch ref/commit from remote
When contributors submit PRs to Flatcar, the Flatcar version produced from that
branch has it's version string generated using `git describe`. It might look
something like this:

  4165.0.0+nightly-20241125-2100-8-gf5d9435f43

We then extract the hash out of the version string and try to check it out.
This only succeeds if the commit is referenced by a ref in the upstream
repository - so this only works for core developers. For PRs from external
contributors the hash is present in the upstream repository but is not fetched
with the default refs. To make it work switch to
`git fetch`+`git checkout FETCH_HEAD`. This only works for full SHA so this change
depends on that being changed in the script repo at the same time.

We also have no use for the full commit history so add `--depth=1` to the fetch
call.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>